make THEN step for visibility ignore irrelevant invisible elements
when checking that COUNT elements are in/visible do not check the count of all found element but rather the count of those that are visible or invisible. This also allows to combine providing the COUNT with using the NOT option, e.g. `then 10 elements should not be visible`.

(cherry picked from commit 1157405106b928b49cfc46a44eb0258d6d244ad9)